+++ test/nodes/test_result_groups_test.rb
@@ -4,6 +4,94 @@ require 'test_helper'
 
 # case test result grouups tests
 class TestResultGroupsTest < Minitest::Test
+  def test_should_filter_by_missing_required_fetal_medicine_test_result_group
+    parser = Canql::Parser.new('all cases with missing required fetal medicine')
+    assert parser.valid?
+    assert_test_result_group_count parser, 1
+    assert_test_result_group_values parser, 0,
+                                    'exists' => { Canql::EQUALS => false },
+                                    'group' => { Canql::EQUALS => 'fetal_medicine' },
+                                    'required' => { Canql::EQUALS => true }
+    assert_equal({ Canql::EQUALS => 'case' }, parser.meta_data['results.subject'])
+  end
+
+  def test_should_filter_by_missing_fetal_medicine_test_result_group
+    parser = Canql::Parser.new('all cases with missing fetal medicine')
+    assert parser.valid?
+    assert_test_result_group_count parser, 1
+    assert_test_result_group_values parser, 0,
+                                    'exists' => { Canql::EQUALS => false },
+                                    'group' => { Canql::EQUALS => 'fetal_medicine' },
+                                    'required' => { Canql::EQUALS => false }
+    assert_equal({ Canql::EQUALS => 'case' }, parser.meta_data['results.subject'])
+  end
+
+  def test_should_filter_by_missing_required_dating_test_result_group
+    parser = Canql::Parser.new('all cases with missing required dating')
+    assert parser.valid?
+    assert_test_result_group_count parser, 1
+    assert_test_result_group_values parser, 0,
+                                    'exists' => { Canql::EQUALS => false },
+                                    'group' => { Canql::EQUALS => 'dating' },
+                                    'required' => { Canql::EQUALS => true }
+    assert_equal({ Canql::EQUALS => 'case' }, parser.meta_data['results.subject'])
+  end
+
+  def test_should_filter_by_missing_dating_test_result_group
+    parser = Canql::Parser.new('all cases with missing dating')
+    assert parser.valid?
+    assert_test_result_group_count parser, 1
+    assert_test_result_group_values parser, 0,
+                                    'exists' => { Canql::EQUALS => false },
+                                    'group' => { Canql::EQUALS => 'dating' },
+                                    'required' => { Canql::EQUALS => false }
+    assert_equal({ Canql::EQUALS => 'case' }, parser.meta_data['results.subject'])
+  end
+
+  def test_should_filter_by_supplied_required_fetal_medicine_test_result_group
+    parser = Canql::Parser.new('all cases with supplied required fetal medicine')
+    assert parser.valid?
+    assert_test_result_group_count parser, 1
+    assert_test_result_group_values parser, 0,
+                                    'exists' => { Canql::EQUALS => true },
+                                    'group' => { Canql::EQUALS => 'fetal_medicine' },
+                                    'required' => { Canql::EQUALS => true }
+    assert_equal({ Canql::EQUALS => 'case' }, parser.meta_data['results.subject'])
+  end
+
+  def test_should_filter_by_supplied_fetal_medicine_test_result_group
+    parser = Canql::Parser.new('all cases with supplied fetal medicine')
+    assert parser.valid?
+    assert_test_result_group_count parser, 1
+    assert_test_result_group_values parser, 0,
+                                    'exists' => { Canql::EQUALS => true },
+                                    'group' => { Canql::EQUALS => 'fetal_medicine' },
+                                    'required' => { Canql::EQUALS => false }
+    assert_equal({ Canql::EQUALS => 'case' }, parser.meta_data['results.subject'])
+  end
+
+  def test_should_filter_by_supplied_required_dating_test_result_group
+    parser = Canql::Parser.new('all cases with supplied required dating')
+    assert parser.valid?
+    assert_test_result_group_count parser, 1
+    assert_test_result_group_values parser, 0,
+                                    'exists' => { Canql::EQUALS => true },
+                                    'group' => { Canql::EQUALS => 'dating' },
+                                    'required' => { Canql::EQUALS => true }
+    assert_equal({ Canql::EQUALS => 'case' }, parser.meta_data['results.subject'])
+  end
+
+  def test_should_filter_by_supplied_dating_test_result_group
+    parser = Canql::Parser.new('all cases with supplied dating')
+    assert parser.valid?
+    assert_test_result_group_count parser, 1
+    assert_test_result_group_values parser, 0,
+                                    'exists' => { Canql::EQUALS => true },
+                                    'group' => { Canql::EQUALS => 'dating' },
+                                    'required' => { Canql::EQUALS => false }
+    assert_equal({ Canql::EQUALS => 'case' }, parser.meta_data['results.subject'])
+  end
+
   def test_should_filter_by_missing_required_screening_test_result_group
     parser = Canql::Parser.new('all cases with missing required screening')
     assert parser.valid?