+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## [Unreleased]
-*no unreleased changes*
+### Added
+* Added fetal medicine and dating options to the test group filter (#50)
 
 ## 4.4.1 / 2021-01-13
 ### Fixed